Revert "Use current price in profit margin calculations"
This reverts commit 4b8237c9d6a71fda6f74b5198ba9571671326b6b.

After last night's conversation, we decided to return to the first version of
the PR, which will come in a subsequent commit.
