+++ daemon/src/routes_maker.rs
@@ -1,9 +1,10 @@
 use crate::auth::Authenticated;
-use crate::model::cfd::{Order, OrderId, Origin};
+use crate::bitmex_price_feed;
+use crate::maker_cfd;
+use crate::model::cfd::{Cfd, Order, OrderId, Origin};
 use crate::model::{Usd, WalletInfo};
 use crate::routes::EmbeddedFileExt;
 use crate::to_sse_event::ToSseEvent;
-use crate::{bitmex_price_feed, cfd_feed, maker_cfd};
 use anyhow::Result;
 use rocket::http::{ContentType, Header, Status};
 use rocket::response::stream::EventStream;
@@ -20,7 +21,7 @@ use xtra::Address;
 
 #[rocket::get("/feed")]
 pub async fn maker_feed(
-    rx_cfds: &State<watch::Receiver<Vec<cfd_feed::Cfd>>>,
+    rx_cfds: &State<watch::Receiver<Vec<Cfd>>>,
     rx_order: &State<watch::Receiver<Option<Order>>>,
     rx_wallet: &State<watch::Receiver<WalletInfo>>,
     rx_quote: &State<watch::Receiver<bitmex_price_feed::Quote>>,
@@ -38,11 +39,11 @@ pub async fn maker_feed(
         let order = rx_order.borrow().clone();
         yield order.to_sse_event();
 
+        let cfds = rx_cfds.borrow().clone();
+        yield cfds.to_sse_event();
         let quote = rx_quote.borrow().clone();
         yield quote.to_sse_event();
 
-        let cfds = rx_cfds.borrow().clone();
-        yield cfds.to_sse_event();
 
         loop{
             select! {
@@ -57,7 +58,6 @@ pub async fn maker_feed(
                 Ok(()) = rx_cfds.changed() => {
                     let cfds = rx_cfds.borrow().clone();
                     yield cfds.to_sse_event();
-                }
                 Ok(()) = rx_quote.changed() => {
                     let quote = rx_quote.borrow().clone();
                     yield quote.to_sse_event();
