+++ daemon/src/taker.rs
@@ -1,4 +1,3 @@
-use crate::cfd_feed::CfdFeed;
 use crate::db::load_all_cfds;
 use crate::model::WalletInfo;
 use crate::wallet::Wallet;
@@ -7,7 +6,7 @@ use bdk::bitcoin::secp256k1::{schnorrsig, SECP256K1};
 use bdk::bitcoin::Network;
 use clap::Clap;
 use futures::StreamExt;
-use model::cfd::Order;
+use model::cfd::{Cfd, Order};
 use rocket::fairing::AdHoc;
 use rocket_db_pools::Database;
 use seed::Seed;
@@ -24,7 +23,6 @@ use xtra::Actor;
 
 mod actors;
 mod bitmex_price_feed;
-mod cfd_feed;
 mod cleanup;
 mod db;
 mod keypair;
@@ -104,7 +102,7 @@ async fn main() -> Result<()> {
 
     let oracle = schnorrsig::KeyPair::new(SECP256K1, &mut rand::thread_rng()); // TODO: Fetch oracle public key from oracle.
 
-    let (cfd_feed_updater, cfd_feed_receiver) = CfdFeed::new(cfd_feed::Role::Taker);
+    let (cfd_feed_sender, cfd_feed_receiver) = watch::channel::<Vec<Cfd>>(vec![]);
     let (order_feed_sender, order_feed_receiver) = watch::channel::<Option<Order>>(None);
     let (wallet_feed_sender, wallet_feed_receiver) = watch::channel::<WalletInfo>(wallet_info);
 
@@ -121,7 +119,7 @@ async fn main() -> Result<()> {
         }
     };
 
-    let (task, mut quote_updates) = bitmex_price_feed::new().await?;
+    let (task, quote_updates) = bitmex_price_feed::new().await?;
     tokio::spawn(task);
 
     let figment = rocket::Config::figment()
@@ -132,7 +130,7 @@ async fn main() -> Result<()> {
         .manage(cfd_feed_receiver)
         .manage(order_feed_receiver)
         .manage(wallet_feed_receiver)
-        .manage(quote_updates.clone())
+        .manage(quote_updates)
         .attach(Db::init())
         .attach(AdHoc::try_on_ignite(
             "SQL migrations",
@@ -170,7 +168,7 @@ async fn main() -> Result<()> {
                     db.clone(),
                     wallet.clone(),
                     schnorrsig::PublicKey::from_keypair(SECP256K1, &oracle),
-                    cfd_feed_updater,
+                    cfd_feed_sender,
                     order_feed_sender,
                     send_to_maker,
                     monitor_actor_address,
@@ -191,27 +189,6 @@ async fn main() -> Result<()> {
                     ),
                 );
                 tokio::spawn(wallet_sync::new(wallet, wallet_feed_sender));
-                tokio::spawn({
-                    let cfd_actor_inbox = cfd_actor_inbox.clone();
-
-                    async move {
-                    loop {
-                        let quote  = quote_updates.borrow().clone();
-
-                        if cfd_actor_inbox.do_send_async(taker_cfd::PriceUpdate(quote)).await.is_err() {
-                            tracing::warn!(
-                                "Could not communicate with the message handler, stopping price feed sync"
-                            );
-                            return;
-                        }
-
-                        if quote_updates.changed().await.is_err() {
-                            tracing::warn!("BitMex price feed receiver not available, stopping price feed sync");
-                            return;
-                        }
-                    }
-                }
-                });
 
                 Ok(rocket.manage(cfd_actor_inbox))
             },
